Move video consent function to its own file
